1.修复了U盘模式选择单个iso文件卡死的问题。 2.新增应用冻结、数据清楚、后台应用定时终止功能。 3.提升compileSdk为35。 4.优化功能体验。 5.修改版本号为1.2.5

+++ app/src/main/java/com/easymanager/core/utils/FileUtils.java
@@ -150,22 +150,25 @@ public class FileUtils {
     public void deleteFile(String dirPath){
         File file = new File(dirPath);
         //判断文件不为null或文件目录存在
-        if (file == null || !file.exists()){
-            System.out.println("file delete error");
-        }
-        //取得这个目录下的所有子文件对象
-        File[] files = file.listFiles();
-        //遍历该目录下的文件对象
-        for (File f: files){
-            //判断子目录是否存在子目录,如果是文件则删除
-            if (f.isDirectory()){
-                deleteFile(f.toString());
-            }else {
-                f.delete();
+        if (file != null && file.exists()){
+            //取得这个目录下的所有子文件对象
+            File[] files = file.listFiles();
+            if(files != null && files.length > 0){
+                //遍历该目录下的文件对象
+                for (File f: files){
+                    //判断子目录是否存在子目录,如果是文件则删除
+                    if (f.isDirectory()){
+                        deleteFile(f.toString());
+                    }else {
+                        f.delete();
+                    }
+                }
             }
+
+            //删除空文件夹  for循环已经把上一层节点的目录清空。
+            file.delete();
         }
-        //删除空文件夹  for循环已经把上一层节点的目录清空。
-        file.delete();
+
     }
 
 }
